backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
torch>=2.2.0
torchvision>=0.17.0
numpy>=1.26.3
Pillow>=10.2.0
python-dotenv>=1.0.1
pydantic>=2.6.0
httpx>=0.26.0
reportlab>=4.0.0
motor>=3.3.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
google-auth>=2.27.0
pydantic-settings>=2.1.0
email-validator>=2.0.0
